Addapts tests to match 7031b552f7a03853187f7e9c62f612e2367c6f1b. Improves DBManager tests.
